#9318 : Refactoring unit test classes so the java packages point to "dotcms" instead of "dotmarketing".
