+++ test/com/dotcms/util/CollectionsUtilsTest.java
@@ -1,8 +1,4 @@
-package com.dotmarketing.util;
-
-import org.junit.Test;
-
-import java.util.*;
+package com.dotcms.util;
 
 import static com.dotcms.util.CollectionsUtils.list;
 import static com.dotcms.util.CollectionsUtils.map;
@@ -10,6 +6,12 @@ import static com.dotcms.util.CollectionsUtils.set;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import org.junit.Test;
 
 /**
  * CollectionsUtils unit test.
